Transparent support for MPI_IN_PLACE
Using the same bigarray as both send and receive buffer in calls to
reduce_bigarray, allreduce_bigarray, and scan_bigarray fails (at least
with openmpi 3.1.3) with

Fatal error: exception Mpi.Error("MPI_ERR_BUFFER: invalid buffer pointer")

This patch checks the buffers and substitutes MPI_IN_PLACE for the send
buffer if necessary. It has the advantage that changes to the function
interfaces are not required; one simply passes the same argument twice.